INTELLIJ-53 Test for resolving packages and renaming support
