+++ testdata/com/liferay/ide/idea/bnd/parser/BasePackageParserTest/resolvePackage/bnd.bnd
@@ -1,0 +1,1 @@
+Import-Package: com.liferay.<caret>test
