+++ testdata/com/liferay/ide/idea/bnd/psi/impl/BndHeaderValuePartManipulatorTest/com/liferay/test/Bar.java
@@ -1,0 +1,5 @@
+package com.liferay.<caret>test;
+
+public class Bar {
+
+}
