+++ testdata/com/liferay/ide/idea/bnd/psi/impl/BndHeaderValuePartManipulatorTest/com/liferay/test/Foo.java
@@ -1,0 +1,5 @@
+package com.liferay.test;
+
+public class Foo {
+
+}
