+++ testdata/com/liferay/ide/idea/bnd/psi/impl/BndHeaderValuePartManipulatorTest/com/liferay/test/Foo_Renamed.java
@@ -1,0 +1,5 @@
+package com.liferay.renamed;
+
+public class Foo {
+
+}
