+++ testdata/com/liferay/ide/idea/bnd/psi/impl/BndHeaderValuePartManipulatorTest/testRenamePackageInsideBnd/bnd.bnd
@@ -1,0 +1,1 @@
+Import-Package: com.liferay.<caret>test
