+++ testdata/com/liferay/ide/idea/bnd/psi/impl/BndHeaderValuePartManipulatorTest/testRenamePackageInsideClass/bnd_renamed.bnd
@@ -1,0 +1,1 @@
+Import-Package: com.liferay.renamed
